Fix problem with databases.

+++ app/src/main/java/com/simplegames/finance/BL/Model/OperationItem.java
@@ -6,6 +6,6 @@ import java.math.BigDecimal;
  * Created by andrey.kakin on 31.10.2014.
  */
 public class OperationItem extends BLBaseModel{
-    public BigDecimal Price;
+    public double Price;
     public Product Product;
 }
