+++ app/src/main/java/com/simplegames/finance/dal/operationItem/OperationItem.java
@@ -1,7 +1,5 @@
 package com.simplegames.finance.dal.operationItem;
 
-import java.math.BigDecimal;
-
 /**
  * Created by andrey.kakin on 13.10.2014.
  */
@@ -10,5 +8,5 @@ public class OperationItem
     public long Id;
     public long OperationId;
     public long ProductId;
-    public BigDecimal Price;
+    public double Price;
 }
